Incorporated PR comments
